backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
fastapi==0.135.2
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.2
idna==3.11
Jinja2==3.1.6
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.1
polars==1.39.3
polars-runtime-32==1.39.3
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.32.5
scipy==1.17.1
setuptools==81.0.0
six==1.17.0
starlette==1.0.0
sympy==1.14.0
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.26
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.42.0
